refactor(compensation): move Failed components to features directory
- Move Failed components from components directory to features directory
- Update import paths for GlobalDrawer and other components
- Rename FailedTable to FailedTable.tsx in import statements
- Update routes to use new FailedView path
