+++ compensation/dashboard/src/features/Failed/__tests__/ErrorDetails.test.tsx
