+++ compensation/dashboard/src/features/Failed/details/ErrorDetails.tsx
