+++ compensation/dashboard/src/features/Failed/details/FailedDetails.tsx
