+++ compensation/dashboard/src/features/Failed/details/FetchingFailedDetails.tsx
